Move ControlFlowCheck to new file
