+++ riscv_analysis/src/lints/mod.rs
@@ -6,3 +6,6 @@ pub use instruction_in_text::*;
 
 mod overlapping_function;
 pub use overlapping_function::*;
+
+mod control_flow;
+pub use control_flow::*;
